Second version of the filter
