+++ libs/utils/Cargo.toml
@@ -11,6 +11,7 @@ path = "lib.rs"
 [dependencies]
 log = "*"
 regex = "0.2.2"
+proxy-wasm = { version = "0.1.*", features = ["wee_alloc"] }
 indexmap = { version = "1.6.1", features = ["serde-1"] }
 petgraph = {version =  "0.*", features = ["serde-1"] }
 serde = { version = "1.0", features = ["derive"] }
